+++ src/Client/Upload.tsx
@@ -58,6 +58,12 @@ export class UploadLink extends ApolloLink {
       }));
     }
 
+    operation.setContext(({ headers }) => {
+      console.log(headers);
+      
+      return {}
+    });
+
     // operation.setContext(({ headers }) => ({
     //   headers: {
     //   }
